Renamed Assert.. to ..Assertion, using noun instead of verb for class name
