+++ test_integration/02_stake_and_mint/utils/progress_stake_assertion.js
@@ -21,9 +21,9 @@
 const assert = require('assert');
 
 /**
- *  Class to assert event and balances after stake
+ *  Class to assert event and balances after progress stake
  */
-class AssertStake {
+class ProgressStakeAssertion {
     /**
      * @param gateway Truffle gateway instance.
      * @param token Truffle token instance.
@@ -36,7 +36,7 @@ class AssertStake {
     }
 
     /**
-     *  This verifies event and balances.
+     *  This verifies event and balances of staker, gateway and stakeVault.
      *
      * @param event Event object after decoding.
      * @param stakeRequest Stake request. {amount: *, gasPrice: *,gasLimit:*
@@ -48,32 +48,36 @@ class AssertStake {
     async verify(event, stakeRequest, initialBalances) {
         await this._assertBalancesForStake(stakeRequest, initialBalances);
 
-        AssertStake._assertStakeEvent(event, stakeRequest);
+        ProgressStakeAssertion._assertProgressStakeEvent(event, stakeRequest);
     }
 
     /**
-     * This capture basetoken and token balance of gateway and staker.
+     * This capture base token and token balance of gateway and staker.
      *
      * @param staker
      * @return {Promise<{baseToken: {gateway: *, staker: *}, token: {gateway: *, staker: *}}>}
      */
     async captureBalances(staker) {
+        const stakeVault = await this.gateway.stakeVault();
         return {
             baseToken: {
                 gateway: await this.baseToken.balanceOf(this.gateway.address),
                 staker: await this.baseToken.balanceOf(staker),
+                stakeVault: await this.baseToken.balanceOf(stakeVault),
             },
             token: {
                 gateway: await this.token.balanceOf(this.gateway.address),
                 staker: await this.token.balanceOf(staker),
+                stakeVault: await this.token.balanceOf(stakeVault),
             },
         };
     }
 
     /**
-     * This asserts balances of staker and gateway  after stake.
+     * This asserts balances of staker, gateway and stakeVault  after
+     * progress stake.
      * @param stakeRequest
-     * @param initialBalances Initial balance of staker and gateway
+     * @param initialBalances Initial balance of staker, gateway and stake vault
      *                        generated by captureBalances method/
      * @private
      */
@@ -82,7 +86,7 @@ class AssertStake {
 
         // Assert gateway balance
         const expectedGatewayBaseTokenBalance = initialBalances.baseToken.gateway
-            .add(stakeRequest.bounty);
+            .sub(stakeRequest.bounty);
 
         // Assert bounty is transferred to gateway.
         assert.strictEqual(
@@ -93,7 +97,7 @@ class AssertStake {
         );
 
         const expectedGatewayTokenBalance = initialBalances.token.gateway
-            .add(stakeRequest.amount);
+            .sub(stakeRequest.amount);
 
         // Assert stake amount is transferred to gateway.
         assert.strictEqual(
@@ -105,7 +109,7 @@ class AssertStake {
 
         // Assert staker balance
         const expectedStakerBaseTokenBalance = initialBalances.baseToken.staker
-            .sub(stakeRequest.bounty);
+            .add(stakeRequest.bounty);
 
         // Assert bounty is transferred to gateway.
         assert.strictEqual(
@@ -115,8 +119,7 @@ class AssertStake {
            instead of ${finalBalances.baseToken.staker.toString(10)}`,
         );
 
-        const expectedStakerTokenBalance = initialBalances.token.staker
-            .sub(stakeRequest.amount);
+        const expectedStakerTokenBalance = initialBalances.token.staker;
 
         // Assert stake amount is transferred from staker.
         assert.strictEqual(
@@ -124,6 +127,27 @@ class AssertStake {
             true,
             `Staker token balance must be ${expectedStakerTokenBalance.toString(10)}
            instead of ${finalBalances.token.staker.toString(10)}`,
+        );
+
+        // Assert stake valult balance.
+
+        const expectedStakeVaultTokenBalance = initialBalances.token.stakeVault
+            .add(stakeRequest.amount);
+
+        assert.strictEqual(
+            expectedStakeVaultTokenBalance.eq(finalBalances.token.stakeVault),
+            true,
+            `Stake value token balance must be ${expectedStakeVaultTokenBalance} instead
+          of ${finalBalances.token.stakeVault}`,
+        );
+
+        const expectedStakeVaultBaseTokenBalance = initialBalances.baseToken.stakeVault;
+
+        assert.strictEqual(
+            expectedStakeVaultBaseTokenBalance.eq(finalBalances.baseToken.stakeVault),
+            true,
+            `Stake vault base token balance must be ${expectedStakeVaultTokenBalance} instead
+          of ${finalBalances.baseToken.stakeVault}`,
         );
     }
 
@@ -133,36 +157,40 @@ class AssertStake {
      * @param stakeRequest
      * @private
      */
-    static _assertStakeEvent(event, stakeRequest) {
-        const eventData = event.StakeIntentDeclared;
+    static _assertProgressStakeEvent(event, stakeRequest) {
+        const eventData = event.StakeProgressed;
 
+        assert.strictEqual(
+            eventData._messageHash,
+            stakeRequest.messageHash,
+            'Message hash must match.',
+        );
         assert.strictEqual(
             eventData._staker,
             stakeRequest.staker,
-            `Staker address from event ${eventData._staker} must be equal to ${stakeRequest.staker}.`,
+            'Staker address must match.',
         );
-
         assert.strictEqual(
-            stakeRequest.nonce.eq(eventData._stakerNonce),
+            eventData._stakerNonce.eq(stakeRequest.nonce),
             true,
-            `Staker nonce from event ${eventData._stakerNonce} 
-            must be equal to ${stakeRequest.nonce.toString(10)}.`,
+            'Staker nonce must match.',
         );
-
         assert.strictEqual(
-            eventData._beneficiary,
-            stakeRequest.beneficiary,
-            `Beneficiary address from event ${eventData._beneficiary} 
-            must be equal to ${stakeRequest.beneficiary}.`,
-        );
-
-        assert.strictEqual(
-            stakeRequest.amount.eq(eventData._amount),
+            eventData._amount.eq(stakeRequest.amount),
             true,
-            `Amount from event ${eventData._amount} must be equal 
-            to ${stakeRequest.amount.toString(10)}.`,
+            'Stake amount must match.',
+        );
+        assert.strictEqual(
+            eventData._proofProgress,
+            false,
+            'Proof progress flag should be false.',
+        );
+        assert.strictEqual(
+            eventData._unlockSecret,
+            stakeRequest.unlockSecret,
+            'Unlock secret must match.',
         );
     }
 }
 
-module.exports = AssertStake;
+module.exports = ProgressStakeAssertion;
