+++ test_integration/02_stake_and_mint/utils/stake_assertion.js
@@ -21,9 +21,9 @@
 const assert = require('assert');
 
 /**
- *  Class to assert event and balances after progress stake
+ *  Class to assert event and balances after stake
  */
-class AssertProgressStake {
+class StakeAssertion {
     /**
      * @param gateway Truffle gateway instance.
      * @param token Truffle token instance.
@@ -36,7 +36,7 @@ class AssertProgressStake {
     }
 
     /**
-     *  This verifies event and balances of staker, gateway and stakeVault.
+     *  This verifies event and balances.
      *
      * @param event Event object after decoding.
      * @param stakeRequest Stake request. {amount: *, gasPrice: *,gasLimit:*
@@ -48,36 +48,32 @@ class AssertProgressStake {
     async verify(event, stakeRequest, initialBalances) {
         await this._assertBalancesForStake(stakeRequest, initialBalances);
 
-        AssertProgressStake._assertProgressStakeEvent(event, stakeRequest);
+        StakeAssertion._assertStakeEvent(event, stakeRequest);
     }
 
     /**
-     * This capture base token and token balance of gateway and staker.
+     * This capture basetoken and token balance of gateway and staker.
      *
      * @param staker
      * @return {Promise<{baseToken: {gateway: *, staker: *}, token: {gateway: *, staker: *}}>}
      */
     async captureBalances(staker) {
-        const stakeVault = await this.gateway.stakeVault();
         return {
             baseToken: {
                 gateway: await this.baseToken.balanceOf(this.gateway.address),
                 staker: await this.baseToken.balanceOf(staker),
-                stakeVault: await this.baseToken.balanceOf(stakeVault),
             },
             token: {
                 gateway: await this.token.balanceOf(this.gateway.address),
                 staker: await this.token.balanceOf(staker),
-                stakeVault: await this.token.balanceOf(stakeVault),
             },
         };
     }
 
     /**
-     * This asserts balances of staker, gateway and stakeVault  after
-     * progress stake.
+     * This asserts balances of staker and gateway  after stake.
      * @param stakeRequest
-     * @param initialBalances Initial balance of staker, gateway and stake vault
+     * @param initialBalances Initial balance of staker and gateway
      *                        generated by captureBalances method/
      * @private
      */
@@ -86,7 +82,7 @@ class AssertProgressStake {
 
         // Assert gateway balance
         const expectedGatewayBaseTokenBalance = initialBalances.baseToken.gateway
-            .sub(stakeRequest.bounty);
+            .add(stakeRequest.bounty);
 
         // Assert bounty is transferred to gateway.
         assert.strictEqual(
@@ -97,7 +93,7 @@ class AssertProgressStake {
         );
 
         const expectedGatewayTokenBalance = initialBalances.token.gateway
-            .sub(stakeRequest.amount);
+            .add(stakeRequest.amount);
 
         // Assert stake amount is transferred to gateway.
         assert.strictEqual(
@@ -109,7 +105,7 @@ class AssertProgressStake {
 
         // Assert staker balance
         const expectedStakerBaseTokenBalance = initialBalances.baseToken.staker
-            .add(stakeRequest.bounty);
+            .sub(stakeRequest.bounty);
 
         // Assert bounty is transferred to gateway.
         assert.strictEqual(
@@ -119,7 +115,8 @@ class AssertProgressStake {
            instead of ${finalBalances.baseToken.staker.toString(10)}`,
         );
 
-        const expectedStakerTokenBalance = initialBalances.token.staker;
+        const expectedStakerTokenBalance = initialBalances.token.staker
+            .sub(stakeRequest.amount);
 
         // Assert stake amount is transferred from staker.
         assert.strictEqual(
@@ -127,27 +124,6 @@ class AssertProgressStake {
             true,
             `Staker token balance must be ${expectedStakerTokenBalance.toString(10)}
            instead of ${finalBalances.token.staker.toString(10)}`,
-        );
-
-        // Assert stake valult balance.
-
-        const expectedStakeVaultTokenBalance = initialBalances.token.stakeVault
-            .add(stakeRequest.amount);
-
-        assert.strictEqual(
-            expectedStakeVaultTokenBalance.eq(finalBalances.token.stakeVault),
-            true,
-            `Stake value token balance must be ${expectedStakeVaultTokenBalance} instead
-          of ${finalBalances.token.stakeVault}`,
-        );
-
-        const expectedStakeVaultBaseTokenBalance = initialBalances.baseToken.stakeVault;
-
-        assert.strictEqual(
-            expectedStakeVaultBaseTokenBalance.eq(finalBalances.baseToken.stakeVault),
-            true,
-            `Stake vault base token balance must be ${expectedStakeVaultTokenBalance} instead
-          of ${finalBalances.baseToken.stakeVault}`,
         );
     }
 
@@ -157,40 +133,36 @@ class AssertProgressStake {
      * @param stakeRequest
      * @private
      */
-    static _assertProgressStakeEvent(event, stakeRequest) {
-        const eventData = event.StakeProgressed;
+    static _assertStakeEvent(event, stakeRequest) {
+        const eventData = event.StakeIntentDeclared;
 
-        assert.strictEqual(
-            eventData._messageHash,
-            stakeRequest.messageHash,
-            'Message hash must match.',
-        );
         assert.strictEqual(
             eventData._staker,
             stakeRequest.staker,
-            'Staker address must match.',
+            `Staker address from event ${eventData._staker} must be equal to ${stakeRequest.staker}.`,
         );
+
         assert.strictEqual(
-            eventData._stakerNonce.eq(stakeRequest.nonce),
+            stakeRequest.nonce.eq(eventData._stakerNonce),
             true,
-            'Staker nonce must match.',
+            `Staker nonce from event ${eventData._stakerNonce} 
+            must be equal to ${stakeRequest.nonce.toString(10)}.`,
         );
+
         assert.strictEqual(
-            eventData._amount.eq(stakeRequest.amount),
+            eventData._beneficiary,
+            stakeRequest.beneficiary,
+            `Beneficiary address from event ${eventData._beneficiary} 
+            must be equal to ${stakeRequest.beneficiary}.`,
+        );
+
+        assert.strictEqual(
+            stakeRequest.amount.eq(eventData._amount),
             true,
-            'Stake amount must match.',
-        );
-        assert.strictEqual(
-            eventData._proofProgress,
-            false,
-            'Proof progress flag should be false.',
-        );
-        assert.strictEqual(
-            eventData._unlockSecret,
-            stakeRequest.unlockSecret,
-            'Unlock secret must match.',
+            `Amount from event ${eventData._amount} must be equal 
+            to ${stakeRequest.amount.toString(10)}.`,
         );
     }
 }
 
-module.exports = AssertProgressStake;
+module.exports = StakeAssertion;
